requirements.txt
# streamlit関連
streamlit
streamlit-aggrid
pandas

# LangChain
langchain
langchain-openai
langchain_community
faiss-cpu
beautifulsoup4
pypdf
pycryptodome

# Gcloud
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
